+++ lib/playback/webcodecs/worklet/tsconfig.json
@@ -11,7 +11,4 @@
 			"path": "../../../common"
 		}
 	],
-	"paths": {
-		"@/*": ["../../*"]
-	}
 }
